refactor(shapes): extract parser logic into focused submodules
Move node-expression, SPARQL-function, and SPARQL-target-type parsing

out of the monolithic crates/shapes/src/shapes.rs and into focused

shapes::parser::{functions, node_expr, target_types} sub-modules.

This reduces shapes.rs by roughly 1,300 lines while keeping the public

API and all existing behavior unchanged.

Refs: #30
