+++ crates/shapes/src/shapes/parser/mod.rs
@@ -1,0 +1,8 @@
+// SPDX-FileCopyrightText: 2026 Blackcat Informatics® Inc. <paudley@blackcatinformatics.ca>
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+//! Focused sub-parsers for the SHACL shapes graph loader.
+
+pub(crate) mod functions;
+pub(crate) mod node_expr;
+pub(crate) mod target_types;
